# add string precision to @printf