closes https://github.com/JuliaLang/julia/issues/14908

thanks to Scott Jones for the solution:

> Date: Tue, 20 Dec 2016 11:49:35 -0500
> From: Scott Jones 
> To: arthurb
> CC: Milan Bouchet-Valat
> Subject: Re: [JuliaLang/julia] @printf string precision truncate (#14908)
> X-Mailer: Apple Mail (2.3256)
> 
> Unfortunately, it's not really that straightforward.
> 
> [1:precision], on a String(), will give you a bounds error if precision is > sizeof the string.
> It won't limit it by the number of characters either, because String is indexed by bytes, not character positions.
> Also, it won't help in preventing buffer overruns (which is the rationale behind the C spec), because Julia will output all of a character if the end range is anywhere within the character
> (so 1:11 may get 13 bytes, for example).
> 
> Here is a function that will correctly limit the string to the number of bytes given with precision, as per C spec.
> (as you can see, it doesn't need to iteratively call nextind either, it takes advantage of the fact that String is in UTF-8, and you can (relatively) efficiently go backwards with prevind)
> 
> function _limit(s, prec)
>            prec >= sizeof(s) && return s
>            p = prevind(s, prec+1)
>            n = nextind(s, p)-1
>            s[1:(prec>=n?n:prevind(s,p))]
> end
> 
> I hope this helps!
> 
> Scott Jones
>