@JeffBezanson 2017-01-05
Thanks, great to have this working!