@bjarthur 2017-01-06
good point.  i've made some changes.  benchmarks folllow.

master:
```
julia> @benchmark @sprintf("%.3s","test")
BenchmarkTools.Trial: 
  memory estimate:  288.00 bytes
  allocs estimate:  5
  --------------
  minimum time:     176.067 ns (0.00% GC)
  median time:      184.563 ns (0.00% GC)
  mean time:        253.849 ns (18.98% GC)
  maximum time:     9.566 μs (88.01% GC)
  --------------
  samples:          10000
  evals/sample:     719
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark @sprintf("%8.3s","test")
BenchmarkTools.Trial: 
  memory estimate:  320.00 bytes
  allocs estimate:  6
  --------------
  minimum time:     362.835 ns (0.00% GC)
  median time:      380.165 ns (0.00% GC)
  mean time:        476.194 ns (12.34% GC)
  maximum time:     34.007 μs (96.67% GC)
  --------------
  samples:          10000
  evals/sample:     206
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

this PR:
```
julia> @benchmark @sprintf("%.3s","test")
BenchmarkTools.Trial: 
  memory estimate:  688.00 bytes
  allocs estimate:  12
  --------------
  minimum time:     368.612 ns (0.00% GC)
  median time:      392.633 ns (0.00% GC)
  mean time:        556.637 ns (21.67% GC)
  maximum time:     33.620 μs (96.52% GC)
  --------------
  samples:          10000
  evals/sample:     206
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark @sprintf("%8.3s","test")
BenchmarkTools.Trial: 
  memory estimate:  432.00 bytes
  allocs estimate:  8
  --------------
  minimum time:     489.338 ns (0.00% GC)
  median time:      501.428 ns (0.00% GC)
  mean time:        655.801 ns (13.01% GC)
  maximum time:     36.257 μs (96.45% GC)
  --------------
  samples:          10000
  evals/sample:     195
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

if precision is not specified the benchmarks are nearly identical so i've omitted them for brevity.